+++ zerver/tests/test_invite.py
@@ -30,7 +30,7 @@ from zerver.actions.create_user import (
     process_new_human_user,
     set_up_streams_for_new_human_user,
 )
-from zerver.actions.default_streams import do_add_default_stream
+from zerver.actions.default_streams import do_add_default_stream, do_remove_default_stream
 from zerver.actions.invites import (
     do_create_multiuse_invite_link,
     do_get_invites_controlled_by_user,
@@ -47,7 +47,10 @@ from zerver.actions.user_settings import do_change_full_name
 from zerver.actions.users import change_user_is_active
 from zerver.context_processors import common_context
 from zerver.lib.create_user import create_user
-from zerver.lib.default_streams import get_default_streams_for_realm_as_dicts
+from zerver.lib.default_streams import (
+    get_default_streams_for_realm_as_dicts,
+    get_slim_realm_default_streams,
+)
 from zerver.lib.send_email import FromAddress, deliver_scheduled_emails, send_future_email
 from zerver.lib.streams import ensure_stream
 from zerver.lib.test_classes import ZulipTestCase
@@ -175,6 +178,7 @@ class InviteUserBase(ZulipTestCase):
         invite_expires_in_minutes: Optional[int] = INVITATION_LINK_VALIDITY_MINUTES,
         body: str = "",
         invite_as: int = PreregistrationUser.INVITE_AS["MEMBER"],
+        subscribe_to_default_streams: bool = False,
         realm: Optional[Realm] = None,
     ) -> "TestHttpResponse":
         """
@@ -198,6 +202,7 @@ class InviteUserBase(ZulipTestCase):
                 "invite_expires_in_minutes": invite_expires_in,
                 "stream_ids": orjson.dumps(stream_ids).decode(),
                 "invite_as": invite_as,
+                "subscribe_to_default_streams": orjson.dumps(subscribe_to_default_streams).decode(),
             },
             subdomain=realm.string_id if realm else "zulip",
         )
@@ -783,6 +788,14 @@ class InviteUserTest(InviteUserBase):
 
         stream_names = ["Rome", "Scotland", "Venice"]
         streams = [get_stream(stream_name, realm) for stream_name in stream_names]
+
+        # You cannot pass "subscribe_to_default_streams" as "True" if passing
+        # streams to subscribe explicitly.
+        self.assert_json_error(
+            self.invite(invitee, stream_names, subscribe_to_default_streams=True),
+            "Invalid parameters: stream_ids and subscribe_to_default_streams are mutually exclusive",
+        )
+
         self.assert_json_success(self.invite(invitee, stream_names))
         self.assertTrue(find_key_by_email(invitee))
         self.submit_reg_form_for_user(invitee, "password")
@@ -792,13 +805,25 @@ class InviteUserTest(InviteUserBase):
         self.assert_json_success(self.invite(invitee, []))
         self.assertTrue(find_key_by_email(invitee))
 
-        default_streams = get_default_streams_for_realm_as_dicts(realm.id)
-        self.assert_length(default_streams, 1)
-
         self.submit_reg_form_for_user(invitee, "password")
-        # If no streams are provided, user is not subscribed to
-        # default streams as well.
         self.check_user_subscribed_only_to_streams("bob", [])
+
+        verona = get_stream("Verona", realm)
+        denmark = get_stream("Denmark", realm)
+        default_streams = get_slim_realm_default_streams(realm.id)
+        self.assert_length(default_streams, 1)
+        self.assertEqual(default_streams, [verona])
+
+        # Check that user is subscribed to the streams that were set as default
+        # at the time of account creation and not at the time of inviting them.
+        invitee = self.nonreg_email("test")
+        self.assert_json_success(self.invite(invitee, [], subscribe_to_default_streams=True))
+        self.assertTrue(find_key_by_email(invitee))
+
+        do_add_default_stream(denmark)
+        do_remove_default_stream(verona)
+        self.submit_reg_form_for_user(invitee, "password")
+        self.check_user_subscribed_only_to_streams("test", [denmark])
 
     def test_can_invite_others_to_realm(self) -> None:
         def validation_func(user_profile: UserProfile) -> bool:
@@ -1296,9 +1321,30 @@ so we didn't send them an invitation. We did send invitations to everyone else!"
             result, "You do not have permission to subscribe other users to streams."
         )
 
-        result = self.invite(invitee, [])
+        # User will be subscribed to default streams even when
+        # the referrer does not permission to subscribe others.
+        result = self.invite(invitee, [], subscribe_to_default_streams=True)
         self.assert_json_success(result)
         self.check_sent_emails([invitee])
+
+        self.submit_reg_form_for_user(invitee, "password")
+
+        default_streams = get_slim_realm_default_streams(realm.id)
+        self.assert_length(default_streams, 1)
+        self.check_user_subscribed_only_to_streams("alice", default_streams)
+        mail.outbox.pop()
+
+        invitee = self.nonreg_email("newguy")
+        # The invited user will not be subscribed to default streams as well.
+        result = self.invite(invitee, [], subscribe_to_default_streams=False)
+        self.assert_json_success(result)
+        self.check_sent_emails([invitee])
+
+        self.submit_reg_form_for_user(invitee, "password")
+        default_streams = get_slim_realm_default_streams(realm.id)
+        self.assert_length(default_streams, 1)
+
+        self.check_user_subscribed_only_to_streams("newguy", [])
         mail.outbox.pop()
 
         self.login("iago")
@@ -1316,6 +1362,17 @@ so we didn't send them an invitation. We did send invitations to everyone else!"
         result = self.invite(invitee, ["Denmark", "Scotland"])
         self.assert_json_success(result)
         self.check_sent_emails([invitee])
+        mail.outbox.pop()
+
+        invitee = self.nonreg_email("test1")
+        result = self.invite(invitee, [], subscribe_to_default_streams=True)
+        self.assert_json_success(result)
+        self.check_sent_emails([invitee])
+
+        self.submit_reg_form_for_user(invitee, "password")
+        self.check_user_subscribed_only_to_streams(
+            "test1", get_slim_realm_default_streams(realm.id)
+        )
 
     def test_invitation_reminder_email(self) -> None:
         # All users belong to zulip realm
@@ -2271,9 +2328,16 @@ class MultiuseInviteTest(ZulipTestCase):
         self.realm.save()
 
     def generate_multiuse_invite_link(
-        self, streams: Optional[List[Stream]] = None, date_sent: Optional[datetime] = None
+        self,
+        streams: Optional[List[Stream]] = None,
+        date_sent: Optional[datetime] = None,
+        subscribe_to_default_streams: bool = False,
     ) -> str:
-        invite = MultiuseInvite(realm=self.realm, referred_by=self.example_user("iago"))
+        invite = MultiuseInvite(
+            realm=self.realm,
+            referred_by=self.example_user("iago"),
+            subscribe_to_default_streams=subscribe_to_default_streams,
+        )
         invite.save()
 
         if streams is not None:
@@ -2371,9 +2435,11 @@ class MultiuseInviteTest(ZulipTestCase):
         name1 = "newuser"
         name2 = "bob"
         name3 = "alice"
+        name4 = "test"
         email1 = self.nonreg_email(name1)
         email2 = self.nonreg_email(name2)
         email3 = self.nonreg_email(name3)
+        email4 = self.nonreg_email(name4)
 
         stream_names = ["Rome", "Scotland", "Venice"]
         streams = [get_stream(stream_name, self.realm) for stream_name in stream_names]
@@ -2388,11 +2454,20 @@ class MultiuseInviteTest(ZulipTestCase):
         self.check_user_subscribed_only_to_streams(name2, streams)
 
         streams = []
-        invite_link = self.generate_multiuse_invite_link(streams=streams)
+        invite_link = self.generate_multiuse_invite_link(
+            streams=streams, subscribe_to_default_streams=True
+        )
         self.check_user_able_to_register(email3, invite_link)
-        # User is not subscribed to default streams as well.
-        self.assert_length(get_default_streams_for_realm_as_dicts(self.realm.id), 1)
-        self.check_user_subscribed_only_to_streams(name3, [])
+        self.assert_length(get_slim_realm_default_streams(self.realm.id), 1)
+        self.check_user_subscribed_only_to_streams(
+            name3, get_slim_realm_default_streams(self.realm.id)
+        )
+
+        invite_link = self.generate_multiuse_invite_link(
+            streams=streams, subscribe_to_default_streams=False
+        )
+        self.check_user_able_to_register(email4, invite_link)
+        self.check_user_subscribed_only_to_streams(name4, [])
 
     def test_multiuse_link_different_realms(self) -> None:
         """
@@ -2442,6 +2517,19 @@ class MultiuseInviteTest(ZulipTestCase):
             {
                 "stream_ids": orjson.dumps(stream_ids).decode(),
                 "invite_expires_in_minutes": 2 * 24 * 60,
+                "subscribe_to_default_streams": orjson.dumps(True).decode(),
+            },
+        )
+        self.assert_json_error(
+            result,
+            "Invalid parameters: stream_ids and subscribe_to_default_streams are mutually exclusive",
+        )
+
+        result = self.client_post(
+            "/json/invites/multiuse",
+            {
+                "stream_ids": orjson.dumps(stream_ids).decode(),
+                "invite_expires_in_minutes": 2 * 24 * 60,
             },
         )
         invite_link = self.assert_json_success(result)["invite_link"]
@@ -2450,11 +2538,36 @@ class MultiuseInviteTest(ZulipTestCase):
 
         self.login("iago")
         stream_ids = []
+        default_streams = get_slim_realm_default_streams(self.realm.id)
+        verona = get_stream("Verona", self.realm)
+        denmark = get_stream("Denmark", self.realm)
+        self.assertEqual(default_streams, [verona])
+
+        # Check that user is subscribed to the streams that were set as default
+        # at the time of account creation and not at the time of inviting them.
         result = self.client_post(
             "/json/invites/multiuse",
             {
                 "stream_ids": orjson.dumps(stream_ids).decode(),
                 "invite_expires_in_minutes": 2 * 24 * 60,
+                "subscribe_to_default_streams": orjson.dumps(True).decode(),
+            },
+        )
+        invite_link = self.assert_json_success(result)["invite_link"]
+
+        do_add_default_stream(denmark)
+        do_remove_default_stream(verona)
+        self.check_user_able_to_register(self.nonreg_email("test1"), invite_link)
+        self.check_user_subscribed_only_to_streams("test1", [denmark])
+
+        self.login("iago")
+        stream_ids = []
+        result = self.client_post(
+            "/json/invites/multiuse",
+            {
+                "stream_ids": orjson.dumps(stream_ids).decode(),
+                "invite_expires_in_minutes": 2 * 24 * 60,
+                "subscribe_to_default_streams": orjson.dumps(False).decode(),
             },
         )
         invite_link = self.assert_json_success(result)["invite_link"]